Committing current WIP.
Created a PaletteWrap that acts as a HOC for palette objects for color pickers.
onSwatchHover has been moved into PaletteWrap, alongside a new prop called onSwatchRemove. The purpose of this is to hold a function for removing an indexed color from a palette. I think this may be an antipattern, however, since an onChange function still needs to be made for new colors being added. Will likely replace this with an onSwatchUpdate function, that does both.
Palettes contain swatches, and reuse the swatch common component.
Removing a swatch from a Palette should be simple, but I'm not keen on introducing a context menu component as either a package (dependency) or new common component (unrelated to the library). Maybe another helper?
Palettes take can take two arrays for the Swatches, both optional. `palette` exists to reflect the state of the custom palette, holding all user-selected swatches. `presetPalette` replaces `presetColors`, for holding what can either be used to initialize `palette`'s state, or as a isolated row of swatches that are immutable and isolated from the user's custom choices.

#305